# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitoring ST2023",
        "version": "1.0"
    },
    "paths": {
...
        "/api/list_kabs": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "Get List",
                "description": "-",
                "operationId": "list_kabs",
                "requestBody": {
                    "description": "wilayah",
                    "required": false
                },
                "responses": {
                    "default": {
                        "description": "return array model ruta"
                    }
                }
            }
        },
        "/api/list_kecs": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "Get List",
                "description": "-",
                "operationId": "list_kecs",
                "parameters": [
                    {
                        "name": "kab_filter",
                        "in": "query",
                        "description": "kode kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "wilayah",
                    "required": false
                },
                "responses": {
                    "default": {
                        "description": "return array model ruta"
                    }
                }
            }
        },
        "/api/list_desas": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "Get List",
                "description": "-",
                "operationId": "list_desas",
                "parameters": [
                    {
                        "name": "kab_filter",
                        "in": "query",
                        "description": "kode kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "kec_filter",
                        "in": "query",
                        "description": "kode kecamatan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "wilayah",
                    "required": false
                },
                "responses": {
                    "default": {
                        "description": "return array model ruta"
                    }
                }
            }
        },
        "/api/list_sls": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "Get List",
                "description": "-",
                "operationId": "list_sls",
                "parameters": [
                    {
                        "name": "kab_filter",
                        "in": "query",
                        "description": "kode kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "kec_filter",
                        "in": "query",
                        "description": "kode kecamatan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desa_filter",
                        "in": "query",
                        "description": "kode desa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "wilayah",
                    "required": false
                },
                "responses": {
                    "default": {
                        "description": "return array model ruta"
                    }
                }
            }
        }
    }
}